editor: request focus when editor gets key event
This fixes an issue where something is stealing the focus from the editor
after typing in a single character in the editor. Not sure why that's
happening, but this seems like a good idea anyway. If a key event gets
sent to the editor it makes sense that it stays focused after it's
handled.