editor: switch back to terminal cursor
This is so we can get the benefits of having the terminal handle the
cursor. In Kitty this means things like cursor trails (i.e. an animated
cursor) when the cursor moves to a new position and showing the text
underneath the cursor automatically.

A future version of Ghostty will also add at least animated cursors, so
this change should be nice. Hopefully Ghostty will also start displaying
the text underneath the cursor soon.